backend/requirements.txt
accelerate>=0.26.0
flask==2.0.1
flask-cors==3.0.10
torch==2.0.1
torchaudio==2.0.1
transformers==4.30.2
outetts==0.2.3
uroman==1.3.1.1
python-dotenv==1.0.1
numpy==1.26.3
pydub
ffmpeg-python